master - обноавлены зависимости и миграции под mysql

+++ database/migrations/2020_07_24_093208_create_task_label_table.php
@@ -14,9 +14,9 @@ class CreateTaskLabelTable extends Migration
     public function up()
     {
         Schema::create('label_task', function (Blueprint $table) {
-            $table->bigInteger('task_id');
+            $table->unsignedInteger('task_id');
             $table->foreign('task_id')->references('id')->on('tasks');
-            $table->bigInteger('label_id');
+            $table->unsignedInteger('label_id');
             $table->foreign('label_id')->references('id')->on('labels');
             $table->unique(['task_id', 'label_id']);
         });
